+++ app/src/main/java/eu/kanade/domain/source/interactor/GetIncognitoState.kt
@@ -2,14 +2,34 @@ package eu.kanade.domain.source.interactor
 
 import eu.kanade.domain.base.BasePreferences
 import eu.kanade.domain.source.service.SourcePreferences
+import eu.kanade.tachiyomi.extension.ExtensionManager
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.distinctUntilChanged
 
 class GetIncognitoState(
     private val basePreferences: BasePreferences,
     private val sourcePreferences: SourcePreferences,
+    private val extensionManager: ExtensionManager,
 ) {
-    fun isEnabled(sourceId: Long): Boolean {
-        val globalIncognito = basePreferences.incognitoMode().get()
-        val sourceIncognito = sourceId.toString() in sourcePreferences.incognitoSources().get()
-        return globalIncognito || sourceIncognito
+    fun await(sourceId: Long?): Boolean {
+        if (basePreferences.incognitoMode().get()) return true
+        if (sourceId == null) return false
+        val extensionPackage = extensionManager.getExtensionPackage(sourceId) ?: return false
+
+        return extensionPackage in sourcePreferences.incognitoExtensions().get()
+    }
+
+    fun subscribe(sourceId: Long?): Flow<Boolean> {
+        if (sourceId == null) return basePreferences.incognitoMode().changes()
+
+        return combine(
+            basePreferences.incognitoMode().changes(),
+            sourcePreferences.incognitoExtensions().changes(),
+            extensionManager.getExtensionPackageAsFlow(sourceId),
+        ) { incognito, incognitoExtensions, extensionPackage ->
+            incognito || (extensionPackage in incognitoExtensions)
+        }
+            .distinctUntilChanged()
     }
 }