+++ app/src/main/java/eu/kanade/domain/source/interactor/ToggleIncognito.kt
@@ -1,0 +1,14 @@
+package eu.kanade.domain.source.interactor
+
+import eu.kanade.domain.source.service.SourcePreferences
+import tachiyomi.core.common.preference.getAndSet
+
+class ToggleIncognito(
+    private val preferences: SourcePreferences,
+) {
+    fun await(extensions: String, enable: Boolean) {
+        preferences.incognitoExtensions().getAndSet {
+            if (enable) it.plus(extensions) else it.minus(extensions)
+        }
+    }
+}