+++ app/src/main/java/eu/kanade/domain/source/interactor/ToggleSource.kt
@@ -8,17 +8,17 @@ class ToggleSource(
     private val preferences: SourcePreferences,
 ) {
 
-    fun awaitDisable(source: Source, enable: Boolean = isEnabled(source.id)) {
-        awaitDisable(source.id, enable)
+    fun await(source: Source, enable: Boolean = isEnabled(source.id)) {
+        await(source.id, enable)
     }
 
-    fun awaitDisable(sourceId: Long, enable: Boolean = isEnabled(sourceId)) {
+    fun await(sourceId: Long, enable: Boolean = isEnabled(sourceId)) {
         preferences.disabledSources().getAndSet { disabled ->
             if (enable) disabled.minus("$sourceId") else disabled.plus("$sourceId")
         }
     }
 
-    fun awaitDisable(sourceIds: List<Long>, enable: Boolean) {
+    fun await(sourceIds: List<Long>, enable: Boolean) {
         val transformedSourceIds = sourceIds.map { it.toString() }
         preferences.disabledSources().getAndSet { disabled ->
             if (enable) disabled.minus(transformedSourceIds) else disabled.plus(transformedSourceIds)